Phase 2

+++ main.js
@@ -14,15 +14,18 @@ var tasksToCreate = document.querySelector('#tasks-to-create');
 leftSection.addEventListener('click', createTaskEvents);
 titleInput.addEventListener('keyup', enableBtns);
 mainSection.addEventListener('click', updateCardEvents);
+mainSection.addEventListener('focusout', saveCard)
 
 // window.addEventListener('load', startOnLoad(e))
 getFromStorage();
 persistOnLoad();
 enableBtns();
+injectAddListMsg();
 
 function createTaskEvents(e) {
   e.preventDefault();
   if (e.target.id === 'make-list-btn') {
+    removeAddListMsg();
     createList();
     clearAllInputs();
     enableBtns();
@@ -51,12 +54,18 @@ function createTaskEvents(e) {
 
 function updateCardEvents(e) {
   e.preventDefault();
-  if (e.target.id === 'delete-btn') {
+  if (e.target.id === `delete-btn${getIdentifier(e)}`) {
     removeList(e);
+    injectAddListMsg();
   };
 
   if (e.target.closest('#urgent-image')) {
     changeUrgency(e);
+  };
+
+  if (e.target.closest('#task-complete-btn')) {
+    completeTask(e);
+    enableDeleteBtn(e);
   };
 };
 
@@ -66,6 +75,18 @@ function getFromStorage() {
   } else {
     toDoArray = JSON.parse(localStorage.getItem('array')).map(element => new ToDoList(element));
   };
+};
+
+function injectAddListMsg() {
+  if (mainSection.innerHTML === '' || mainSection.innerHTML === ' ') {
+    mainSection.insertAdjacentHTML("afterbegin", 
+      `
+      <article id="add-list-msg">
+        <p>Add a to-do list and get checkin' those boxes!</p>
+      </article>
+      `
+    );
+  }; 
 };
 
 function persistOnLoad() {
@@ -96,47 +117,67 @@ function displayTasks(taskObj) {
 };
 
 function createList(e) {
-    var newList = new ToDoList({title: titleInput.value, tasks: currentTasks, id: Date.now(), urgent: false});
+  var newList = new ToDoList({title: titleInput.value, tasks: currentTasks, id: Date.now(), urgent: false});
 
-    toDoArray.push(newList);
-    newList.saveToStorage(toDoArray);
-    /*sortLists();
-    removeIntro();*/
-    displayList(newList);
+  toDoArray.push(newList);
+  newList.saveToStorage(toDoArray);
+  displayList(newList);
 };
 
 function displayList(toDoObj) {
-  var urgent = 'images/urgent.svg';
+  var urgent;
+  var urgentClass;
+  var urgentIcon;
 
   if (toDoObj.urgent === true) {
     urgent = 'images/urgent-active.svg'
+    urgentClass = 'urgent'
+    urgentIcon = 'urgent-icon'
   } else {
     urgent = 'images/urgent.svg'
+    urgentClass = ''
+    urgentIcon = ''
   };
 
   mainSection.insertAdjacentHTML('afterbegin', 
     `
-    <article class="article" data-identifier="${toDoObj.id}">
-      <h3 class="article__title" contenteditable="true">${toDoObj.title}</h3>
-      <ul class="article__ul" id="card-task-list" contenteditable="true">${makeListItems(toDoObj.tasks)}</ul> 
-      <footer class="article__footer">  
-        <img src="${urgent}" id="urgent-image" alt="lightning bolt icon">
-        <img src="images/delete.svg" id="delete-btn" alt="green x">
+    <article class="article ${urgentClass}" id="card${toDoObj.id}" data-identifier="${toDoObj.id}">
+      <h3 class="article__title ${urgentClass}" contenteditable="true" id="card-title${toDoObj.id}">${toDoObj.title}</h3>
+      <ul class="article__ul ${urgentClass}" id="card-list${toDoObj.id}">${makeListItems(toDoObj.tasks)}</ul> 
+      <footer class="article__footer ${urgentClass}" id="card-footer${toDoObj.id}">  
+        <div>
+          <img src="${urgent}" id="urgent-image" alt="lightning bolt icon">
+          <p class="${urgentIcon}" id="urgent-label${toDoObj.id}">URGENT</p>
+        </div>
+        <div>
+          <img src="images/delete.svg" id="delete-btn${toDoObj.id}" alt="green x">
+          <p id="delete-label${toDoObj.id}">DELETE</p>
+        </div>
       </footer>  
-    `);
+    `
+  );
 };
 
 function makeListItems(taskObj) {
-  var checked;
-  var fontStyle;
   var listItems = '';
 
   taskObj.forEach(function(li) {
+    var imgSource;
+    var taskComplete;
+
+    if (li.complete) {
+      imgSource = 'images/checkbox-active.svg';
+      taskComplete = 'complete'
+    } else {
+      imgSource = 'images/checkbox.svg';
+      taskComplete = ''
+    };
+
     listItems += 
     `
-    <li class="" data-identifier="${li.id}">
-    <img src="images/checkbox.svg" id="task-complete-btn">
-    ${li.task}
+    <li class="article__li ${taskComplete}" data-identifier="${li.id}" data-complete="${li.complete}" contenteditable="true">
+      <img src="${imgSource}" id="task-complete-btn">
+      ${li.task}
     </li>
     `
   });
@@ -148,7 +189,6 @@ function clearTaskInput() {
 };
 
 function clearAllInputs() {
-  console.log('ran clearAllInputs')
   titleInput.value = '';
   taskInput.value = '';
   currentTasks = [];
@@ -181,6 +221,7 @@ function enableClearAllBtn() {
 };
 
 function getIdentifier(e) {
+  // console.log(e.target.closest('article').dataset.identifier);
   return e.target.closest('article').dataset.identifier;
 };
 
@@ -191,9 +232,11 @@ function getIndex(e) {
 };
 
 function removeList(e) {
+  if (e.target.src.includes('delete-active.svg')) {
   e.target.closest('article').remove();
   toDoArray[getIndex(e)].deleteFromStorage(getIdentifier(e));
-  // injectIntro();
+  injectAddListMsg();
+  };
 };
 
 function getTaskIdentifier(e) {
@@ -207,26 +250,83 @@ function getTaskIndex(e) {
 };
 
 function removeTask(e) {
-  // debugger;
-  console.log(currentTasks[getTaskIndex(e)], e.target.closest('li'))
   currentTasks.splice(getTaskIndex(e), 1);
   e.target.closest('li').remove();
-  // return currentTasks;
 };
 
-function changeUrgencyImg(e) {
-  var urgentActive = 'images/urgent-active.svg';
-  var urgentNotActive = 'images/urgent.svg';
+function completeTask(e) {
+  var listID = parseInt(getIdentifier(e));
+  var taskID = parseInt(e.target.closest('li').dataset.identifier);
+  var toDoIndex = toDoArray.findIndex(todo => todo.id === listID);
+  var taskIndex = toDoArray[toDoIndex].tasks.findIndex(task => task.id === taskID);   
 
-  if (toDoArray[getIndex(e)].urgent === true) {
-    e.target.src = urgentActive;
+  toDoArray[toDoIndex].updateTask(toDoArray, taskIndex) 
+  changeTaskStyle(e);  
+};
+
+function changeTaskStyle(e) {
+  if (e.target.src.includes('images/checkbox.svg')) {
+    e.target.src = 'images/checkbox-active.svg';
+    e.target.closest('li').dataset.complete = 'true';
+    e.target.closest('li').classList.add('complete');
   } else {
-    e.target.src = urgentNotActive;
+    e.target.src = 'images/checkbox.svg';
+    e.target.closest('li').dataset.complete = 'false';
+    e.target.closest('li').classList.remove('complete');
   };
 };
 
 function changeUrgency(e) {
-  toDoArray[getIndex(e)].urgent = !toDoArray[getIndex(e)].urgent;
-  toDoArray[getIndex(e)].saveToStorage(toDoArray);
   changeUrgencyImg(e);
+  toDoArray[getIndex(e)].updateToDo(toDoArray);
+
+  if (e.target.src.includes('images/urgent.svg')) {
+    e.target.src = 'images/urgent-active.svg'
+  } else {
+    e.target.src = 'images/urgent.svg'
+  };
+};
+
+function changeUrgencyImg(e) {
+    document.querySelector(`#card${getIdentifier(e)}`).classList.toggle('urgent');
+    document.querySelector(`#card-title${getIdentifier(e)}`).classList.toggle('urgent');
+    document.querySelector(`#card-list${getIdentifier(e)}`).classList.toggle('urgent');
+    document.querySelector(`#card-footer${getIdentifier(e)}`).classList.toggle('urgent');
+    document.querySelector(`#urgent-label${getIdentifier(e)}`).classList.toggle('urgent-icon');
+};
+
+function enableDeleteBtn(e) {
+  var listItems = e.target.parentNode.parentNode.children;
+  var listItemsArray = Array.from(listItems);
+  var completeCount = 0;
+
+  listItemsArray.forEach(function(listItem) {
+    if (JSON.parse(listItem.dataset.complete) === true) {
+      completeCount++;
+    };
+  });
+
+  if (completeCount === listItemsArray.length) {
+    document.querySelector(`#delete-btn${getIdentifier(e)}`).src = 'images/delete-active.svg';
+    document.querySelector(`#delete-label${getIdentifier(e)}`).classList.add('urgent-icon');
+  } else {
+    document.querySelector(`#delete-btn${getIdentifier(e)}`).src = 'images/delete.svg';
+    document.querySelector(`#delete-label${getIdentifier(e)}`).classList.remove('urgent-icon');   
+  };
+};
+
+function removeAddListMsg() {
+  var element = document.getElementById('add-list-msg');
+
+  if (element) {
+    element.parentNode.removeChild(element);
+  };  
+};
+
+function saveCard(e) {
+  if (e.target.closest('.article__title')) {
+    var articleTitle = e.target.closest('.article__title').innerText;
+    toDoArray[getIndex(e)].title = articleTitle;
+    toDoArray[getIndex(e)].saveToStorage(toDoArray);
+  };
 };
